Julio 10 de 2017
Subido por Miguel Angel Manrique Latorre
-Creación de página Usuarios Tipos Sueldos

+++ designer/designer-ejb/src/java/Entidades/UsuariosInforeportes.java
@@ -1,7 +1,7 @@
 package Entidades;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
@@ -18,7 +18,7 @@ public class UsuariosInforeportes implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "SECUENCIA")
-    private BigDecimal secuencia;
+    private BigInteger secuencia;
     @JoinColumn(name = "USUARIO", referencedColumnName = "SECUENCIA")
     @ManyToOne(optional = false)
     private Usuarios usuario;
@@ -29,15 +29,15 @@ public class UsuariosInforeportes implements Serializable {
     public UsuariosInforeportes() {
     }
 
-    public UsuariosInforeportes(BigDecimal secuencia) {
+    public UsuariosInforeportes(BigInteger secuencia) {
         this.secuencia = secuencia;
     }
 
-    public BigDecimal getSecuencia() {
+    public BigInteger getSecuencia() {
         return secuencia;
     }
 
-    public void setSecuencia(BigDecimal secuencia) {
+    public void setSecuencia(BigInteger secuencia) {
         this.secuencia = secuencia;
     }
